+++ changelog.md
@@ -32,6 +32,9 @@
     * It is possible to load topics, audios and videos for a group. In order to do so, you need to go to the group buffer and press the menu key, or right mouse click, in the tree item representing the group. New buffers will be created inside the current group's buffer.
     * You can create or delete all buffers for groups by pressing the menu key or right mouse click in the "communities" buffer.
     * There is  support for group topics. When opening them, they will be displayed as a list of posts. You can like or reply to such posts, as well as adding new posts in the topic.
+* Authentication errors should be handled gracefully by the application:
+    * When there is a password change, Socializer must be reauthorized again. An error message will indicate this if the user forgot to do that. After the error, the app will be restarted, prompting the user to introduce the new data for authorizing the application.
+    * If the user introduced incorrect or invalid data, Socializer will display an error and prompt the user again for valid information.
 
 ## Changes in version 0.18 (21.01.2019)